backend/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
certifi==2024.2.2
distro==1.9.0
dj-database-url==2.1.0
Django==5.0.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
dnspython==2.6.1
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
openai==1.14.2
packaging==24.0
psycopg==3.1.18
psycopg2-binary==2.9.9
pydantic==2.6.4
pydantic_core==2.16.3
pymongo==4.6.3
python-dotenv==1.0.1
pytz==2024.1
setuptools==68.2.2
sniffio==1.3.1
sqlparse==0.4.4
tqdm==4.66.2
typing_extensions==4.10.0
wheel==0.41.2
